attempt fixing thumbs for short vids + fix ip/time for register
The thumbnail fix was tested using a converted gif file of Bingo Heeler. In this 4 second long video, the 3rd thumbnail was seconds behind the 2nd thumbnail.

The register fix fixes an oversight where newly registered users didn't have data relating to IPs and the unix timestamp of when they last connected.