getVideoData returns views and remove countRecommended
getVideoData now actually returns the view counter, which makes things more consistent with the site's alternative ways of getting the view counter.

countRecommended has been removed as it both didn't work (it counted all uploaded videos besides the one being watched) and it's only useful on the 2005 layout, which has been replaced with 2013.